backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.4
pydantic-settings==2.2.1
python-multipart==0.0.9
google-generativeai==0.4.1
python-dotenv==1.0.1
pyyaml==6.0.1
jinja2==3.1.3
xhtml2pdf==0.2.15
anthropic==0.40.0
